[clean] remove constants; separate merge
