+++ src/storage/seg/seg.c
@@ -4,6 +4,7 @@
 #include "hashtable.h"
 #include "item.h"
 #include "segevict.h"
+#include "segmerge.h"
 #include "ttlbucket.h"
 #include "datapool/datapool.h"
 
@@ -39,11 +40,6 @@ bool use_cas = false;
 pthread_t     bg_tid;
 volatile bool stop     = false;
 
-static int64_t n_merge_seg       = 0;
-static int64_t merge_seg_age_sum = 0;
-static int     min_mature_time   = 20;
-
-
 static char *seg_state_change_str[] = {
     "allocated",
     "concurrent_get",
@@ -52,18 +48,10 @@ static char *seg_state_change_str[] = {
     "invalid_reason",
 };
 
-
-static inline void
-prep_seg_to_merge(int32_t start_seg_id, struct seg *segs_to_merge[],
-                  int *n_seg_to_merge, double *merge_keep_ratio);
-int32_t
-merge_segs(struct seg *segs_to_merge[], int at_most_n_seg);
-
-
 /**
  * wait until no other threads are accessing the seg (refcount == 0)
  */
-static void
+void
 seg_wait_refcnt(int32_t seg_id)
 {
     struct seg *seg          = &heap.segs[seg_id];
@@ -115,40 +103,8 @@ seg_is_accessible(int32_t seg_id)
     bool       expired = seg->ttl + seg->create_at < time_proc_sec() \
  || seg->create_at <= flush_at;
 
-#ifdef NO_BACKGROUND_EXPIRATION
-    if (expired)
-        seg_rm_expired_seg(seg_id);
-#endif
-
     return !expired;
 }
-
-//bool
-//seg_r_ref(int32_t seg_id)
-//{
-//    struct seg *seg = &heap.segs[seg_id];
-//
-//    if (seg_is_accessible(seg_id)) {
-////    if (__atomic_load_n(&seg->accessible, __ATOMIC_RELAXED) == 1) {
-//        /* this does not strictly prevent race condition, but it is fine
-//         * because letting one reader passes when the segment is locking
-//         * has no problem in correctness */
-//        __atomic_fetch_add(&seg->r_refcount, 1, __ATOMIC_RELAXED);
-//        return true;
-//    }
-//
-//    return false;
-//}
-
-//void
-//seg_r_deref(int32_t seg_id)
-//{
-//    struct seg *seg = &heap.segs[seg_id];
-//
-//    int16_t ref = __atomic_sub_fetch(&seg->r_refcount, 1, __ATOMIC_RELAXED);
-//
-//    ASSERT(ref >= 0);
-//}
 
 bool
 seg_w_ref(int32_t seg_id)
@@ -184,19 +140,14 @@ seg_w_deref(int32_t seg_id)
  * in order to avoid eviction algorithm picking this seg,
  * we do not clear seg->locked until it is linked into ttl_bucket
  */
-static void
+void
 seg_init(int32_t seg_id)
 {
     ASSERT(seg_id != -1);
     struct seg *seg        = &heap.segs[seg_id];
     uint8_t    *data_start = get_seg_data_start(seg_id);
 
-    /* I think we still need this to prevent incorrect data
-     * when evicting/expiring a segment that is waiting for a super super
-     * long write
-     **/
 //    cc_memset(data_start, 0, heap.seg_size);
-//    cc_memset(data_start + heap.seg_size / 2, 0, heap.seg_size / 2);
 
 #if defined CC_ASSERT_PANIC || defined CC_ASSERT_LOG
     *(uint64_t *) (data_start) = SEG_MAGIC;
@@ -225,7 +176,7 @@ seg_init(int32_t seg_id)
     seg->n_active_byte = 0;
 }
 
-static void
+void
 rm_seg_from_ttl_bucket(int32_t seg_id)
 {
     struct seg        *seg        = &heap.segs[seg_id];
@@ -281,7 +232,7 @@ rm_seg_from_ttl_bucket(int32_t seg_id)
  * because multiple threads could try to evict/expire the seg at the same time,
  * return true if current thread is able to grab the lock, otherwise false
  */
- /* TODO(jason): separate into two func: one lock for remove, one remove */
+/* TODO(jason): separate into two func: one lock for remove, one remove */
 bool
 rm_all_item_on_seg(int32_t seg_id, enum seg_state_change reason)
 {
@@ -414,24 +365,17 @@ rm_all_item_on_seg(int32_t seg_id, enum seg_state_change reason)
 
     /* expensive debug commands */
     if (seg->n_item != 0) {
-        log_warn("removing all items from segment, but %d items left", seg->n_item);
+        log_warn("removing all items from segment, but %d items left",
+            seg->n_item);
         scan_hashtable_find_seg(seg_id);
     }
 
     ASSERT(seg->n_item == 0);
     ASSERT(seg->occupied_size == 0 || seg->occupied_size == 8);
 
-    if (reason == SEG_EXPIRATION) {
-        INCR(seg_metrics, seg_expire);
-    }
-    else if (reason == SEG_EVICTION) {
-        INCR(seg_metrics, seg_evict);
-    } else {
-        ASSERT(0);
-    }
-
     return true;
 }
+
 
 rstatus_i
 expire_seg(int32_t seg_id)
@@ -444,19 +388,22 @@ expire_seg(int32_t seg_id)
     int status = pthread_mutex_lock(&heap.mtx);
     ASSERT(status == 0);
 
-    add_seg_to_freepool(seg_id, SEG_EXPIRATION);
+    seg_add_to_freepool(seg_id, SEG_EXPIRATION);
 
     pthread_mutex_unlock(&heap.mtx);
+
+    INCR(seg_metrics, seg_expire);
 
     return CC_OK;
 }
 
 /**
  * get a seg from free pool,
- * @params use_reserved merge-based eviction reserves one seg per thread
  *
+ * use_reserved: merge-based eviction reserves one seg per thread
+ * return the segment id if there are free segment, -1 if not
  */
-static inline int32_t
+int32_t
 seg_get_from_freepool(bool use_reserved)
 {
     int32_t seg_id_ret, next_seg_id;
@@ -482,6 +429,7 @@ seg_get_from_freepool(bool use_reserved)
 
     seg_id_ret = heap.free_seg_id;
     ASSERT(seg_id_ret >= 0);
+
     next_seg_id = heap.segs[seg_id_ret].next_seg_id;
     heap.free_seg_id = next_seg_id;
     if (next_seg_id != -1) {
@@ -495,219 +443,12 @@ seg_get_from_freepool(bool use_reserved)
     return seg_id_ret;
 }
 
-//#ifdef USE_MERGE
-static bool
-check_merge_seg(void)
-{
-    struct seg *seg  = NULL, *next1_seg, *next2_seg = NULL;
-
-    int32_t           ttl_bkt_idx;
-    struct ttl_bucket *ttl_bkt;
-    int               i;
-    bool       found = false;
-
-    static __thread int32_t    last_ttl_bkt_idx = 0;
-    static __thread struct seg *segs_to_merge[N_MAX_SEG_MERGE];
-    static __thread double     merge_keep_ratio[N_MAX_SEG_MERGE];
-    static __thread int32_t    at_most_n_seg;
-
-    pthread_mutex_lock(&heap.mtx);
-    if (heap.n_free_seg > heap.n_reserved_seg) {
-        pthread_mutex_unlock(&heap.mtx);
-        return true;
-    }
-    pthread_mutex_unlock(&heap.mtx);
-
-    int n_retry = -1;
-    test:
-    n_retry += 1;
-    /* it is important to have MAX_N_TTL_BUCKET+1, because
-     * if there is only one TTL bucket, we need to check this
-     * ttl bucket again after reaching the end of bucket */
-    for (i      = 0; i < MAX_N_TTL_BUCKET + 1; i++) {
-        ttl_bkt_idx = (last_ttl_bkt_idx + i) % MAX_N_TTL_BUCKET;
-        ttl_bkt     = &ttl_buckets[ttl_bkt_idx];
-        if (ttl_buckets[ttl_bkt_idx].first_seg_id == -1) {
-            continue;
-        }
-
-        if (pthread_mutex_trylock(&ttl_bkt->mtx) != 0) {
-            /* with more than 16 threads and 20% write, this lock becomes
-             * the bottleneck, so for scalability, we just check next TTL bucket
-             */
-            continue;
-        }
-
-        if (ttl_bkt->next_seg_to_merge != -1) {
-            seg = &heap.segs[ttl_bkt->next_seg_to_merge];
-        }
-        else {
-            seg = &heap.segs[ttl_bkt->first_seg_id];
-        }
-
-        while (1) {
-            if (seg->next_seg_id == -1) {
-                break;
-            }
-            next1_seg = &heap.segs[seg->next_seg_id];
-
-            if (next1_seg->next_seg_id == -1) {
-                break;
-            }
-            next2_seg = &heap.segs[next1_seg->next_seg_id];
-
-            if (next2_seg->next_seg_id == -1) {
-                break;
-            }
-
-            if (seg_mergeable(seg)) {
-                if (seg_mergeable(next1_seg)) {
-                    if (seg_mergeable(next2_seg)) {
-                        found = true;
-                        break;
-                    }
-                    else {
-                        if (next2_seg->next_seg_id != -1) {
-                            seg = &heap.segs[next2_seg->next_seg_id];
-                            continue;
-                        }
-                        else {
-                            break;
-                        }
-                    }
-                }
-                else {
-                    seg = next2_seg;
-                    continue;
-                }
-            }
-            else {
-                seg = next1_seg;
-                continue;
-            }
-        }
-
-        if (!found) {
-            ttl_buckets[ttl_bkt_idx].next_seg_to_merge = -1;
-            int32_t      seg_id        = ttl_buckets[ttl_bkt_idx].first_seg_id;
-            delta_time_i first_seg_age = time_proc_sec() -
-                heap.segs[seg_id].create_at;
-            /* the segments in this bucket cannot be merged, but it has been
-             * too old, we evict it */
-
-
-            if (n_merge_seg > 100
-                && first_seg_age > (merge_seg_age_sum / n_merge_seg) * 2) {
-                int success = rm_all_item_on_seg(seg_id, SEG_EXPIRATION);
-                if (success) {
-                    pthread_mutex_lock(&heap.mtx);
-                    add_seg_to_freepool(seg_id, SEG_EVICTION);
-                    pthread_mutex_unlock(&heap.mtx);
-                    last_ttl_bkt_idx = ttl_bkt_idx + 1;
-                    pthread_mutex_unlock(&ttl_bkt->mtx);
-                    return true;
-                }
-            }
-
-
-            /* next ttl bucket please */
-            pthread_mutex_unlock(&ttl_bkt->mtx);
-            continue;
-        }
-
-
-        /* block the eviction of next N_MAX_SEG_MERGE segments */
-        prep_seg_to_merge(seg->seg_id, segs_to_merge, &at_most_n_seg,
-            merge_keep_ratio);
-        pthread_mutex_unlock(&ttl_bkt->mtx);
-
-        /* I hope I can move the ttl_bkt lock out of merge_segs
-         * it is not clear when it is in the merge_segs */
-        ttl_buckets[ttl_bkt_idx].next_seg_to_merge =
-            merge_segs(segs_to_merge, at_most_n_seg);
-        last_ttl_bkt_idx = ttl_bkt_idx;
-
-        return true;
-    }
-
-//    for (int j=0; j<heap.max_nseg; j++) {
-//        seg_print_warn(j);
-//        log_warn("%d mergeable %d", j, seg_mergeable(&heap.segs[j]));
-//    }
-    log_warn("cannot find mergeable seg, retry %d", n_retry);
-
-//    for (i = 0; i < MAX_N_TTL_BUCKET+1; i++) {
-//        if (ttl_buckets[i].first_seg_id == -1)
-//            continue;
-//        seg = &heap.segs[ttl_buckets[i].first_seg_id];
-//        while (seg != NULL) {
-//            printf("seg %d (%d), ", seg->seg_id, seg_mergeable(seg));
-//            if (seg->next_seg_id != -1)
-//                seg = &heap.segs[seg->next_seg_id];
-//            else
-//                seg = NULL;
-//        }
-//        printf("\n");
-//    }
-    if (n_retry < 8) {
-        usleep(200 * n_retry * n_retry);
-        min_mature_time = min_mature_time / 2;
-        goto test;
-    }
-
-    char s[64];
-    int32_t j;
-    for (j = 0; j < heap.max_nseg; j++) {
-        snprintf(s, 64, "%d mergeable %d", j, seg_mergeable(&heap.segs[j]));
-        SEG_PRINT(j, s, log_warn);
-    }
-//    for (i = 0; i < MAX_N_TTL_BUCKET+1; i++) {
-//        if (ttl_buckets[i].first_seg_id == -1)
-//            continue;
-//        seg = &heap.segs[ttl_buckets[i].first_seg_id];
-//        while (seg != NULL) {
-//            printf("seg %d (%d), ", seg->seg_id, seg_mergeable(seg));
-//            if (seg->next_seg_id != -1)
-//                seg = &heap.segs[seg->next_seg_id];
-//            else
-//                seg = NULL;
-//        }
-//        printf("\n");
-//    }
-    ASSERT(0);
-    return false;
-}
-
-
-static int32_t
-seg_get_new_with_merge(void)
-{
-//    static proc_time_i last_merge = 0, last_clear = 0;
-//    if (last_merge == 0) {
-//        last_merge = time_proc_sec();
-//        last_clear = time_proc_sec();
-//    }
-
-    int32_t seg_id_ret;
-
-    while ((seg_id_ret = seg_get_from_freepool(false)) == -1) {
-        if (!check_merge_seg()) {
-            /* better evict a random one */
-            return -1;
-        }
-    }
-    seg_init(seg_id_ret);
-
-    return seg_id_ret;
-}
-//#endif
-
 /**
  * add evicted/allocated seg to free pool,
  * caller should grab the heap lock before calling this function
  **/
 void
-add_seg_to_freepool(int32_t seg_id, enum seg_state_change reason)
+seg_add_to_freepool(int32_t seg_id, enum seg_state_change reason)
 {
     ASSERT(pthread_mutex_trylock(&heap.mtx) != 0);
 
@@ -737,486 +478,73 @@ add_seg_to_freepool(int32_t seg_id, enum seg_state_change reason)
             seg_state_change_str[reason], seg_id, heap.n_free_seg);
 }
 
+//static int32_t
+//seg_get_new_using_merge(void)
+//{
+//    int32_t seg_id_ret;
+//
+//    while ((seg_id_ret = seg_get_from_freepool(false)) == -1) {
+//        if (!check_merge_seg()) {
+//            /* better evict a random one */
+//            return -1;
+//        }
+//    }
+//    seg_init(seg_id_ret);
+//
+//    return seg_id_ret;
+//}
+
 /**
- * get a new segment, we search for a free segment in the following order
+ * get a new segment, search for a free segment in the following order
  * 1. unallocated heap
  * 2. free pool
  * 3. eviction
  **/
-static int32_t
-seg_get_new_no_merge(void)
+int32_t
+seg_get_new(void)
 {
+#define MAX_RETRIES 8
     evict_rstatus_e status;
     int32_t         seg_id_ret;
+    /* eviction may fail if other threads pick the same seg */
+    int             n_retries_left = MAX_RETRIES;
 
-    INCR(seg_metrics, seg_req);
+    INCR(seg_metrics, seg_get);
 
-    if ((seg_id_ret = seg_get_from_freepool(true)) != -1) {
-        /* free pool has seg */
-        log_verb("get_new_seg: allocate seg %" PRId32 " from free pool",
-            seg_id_ret);
-    }
-    else {
-        /* evict one seg */
-        int n_evict_retries = 0;
-        while (1) {
-            /* eviction may fail if other threads pick the same seg
-             * (can happen in random eviction) */
-            status = least_valuable_seg(&seg_id_ret);
-            if (status == EVICT_NO_SEALED_SEG) {
-                log_warn("unable to evict seg because no seg can be evicted");
-                INCR(seg_metrics, seg_req_ex);
+    seg_id_ret = seg_get_from_freepool(false);
 
+    while (seg_id_ret == -1 && n_retries_left >= 0) {
+        /* evict seg */
+        if (evict_info.policy == EVICT_MERGE_FIFO) {
+            if (!check_merge_seg()) {
+                /* better evict a random one */
                 return -1;
             }
+            seg_id_ret = seg_get_from_freepool(false);
 
-            if (rm_all_item_on_seg(seg_id_ret, SEG_EVICTION)) {
-                log_verb("get_new_seg: allocate seg %" PRId32 " from eviction",
-                    seg_id_ret);
+        } else {
+            status = seg_evict(&seg_id_ret);
+            if (status == EVICT_OK || status == EVICT_NO_AVAILABLE_SEG) {
                 break;
             }
-
-            if (++n_evict_retries >= 8) {
-                log_error("get_new_seg: unable to evict after 8 retries");
-                return -1;
-            }
         }
+
+        if (n_retries_left-- < MAX_RETRIES) {
+            INCR(seg_metrics, seg_evict_retry);
+        }
+    }
+
+    if (seg_id_ret == -1) {
+        INCR(seg_metrics, seg_get_ex);
+        log_error("unable to get new seg from eviction");
+
+        return -1;
     }
 
     seg_init(seg_id_ret);
 
     return seg_id_ret;
 }
-
-int32_t
-get_new_seg(void)
-{
-    if (evict_info.policy == EVICT_MERGE_FIFO) {
-        return seg_get_new_with_merge();
-    }
-    else {
-        return seg_get_new_no_merge();
-    }
-}
-
-//#ifdef USE_MERGE
-static inline void
-seg_copy(int32_t seg_id_dest, int32_t seg_id_src,
-         double *cutoff_freq, double target_ratio)
-{
-    struct item *it;
-    struct item *last_it       = NULL;
-    struct seg  *seg_dest      = &heap.segs[seg_id_dest];
-    struct seg  *seg_src       = &heap.segs[seg_id_src];
-    uint8_t     *seg_data_src  = get_seg_data_start(seg_id_src);
-    uint8_t     *curr_src      = seg_data_src;
-
-    uint8_t  *seg_data_dest = get_seg_data_start(seg_id_dest);
-    uint32_t offset         =
-                 MIN(seg_src->write_offset, heap.seg_size) - ITEM_HDR_SIZE;
-
-    int32_t it_sz = 0;
-    bool item_up_to_date;
-    bool seg_in_full = false;
-
-#ifdef DUMP_FOR_ANALYSIS
-    FILE *f = fopen("freq.real", "a");
-    FILE *f2 = fopen("freq.APFC", "a");
-#endif
-
-#if defined CC_ASSERT_PANIC || defined CC_ASSERT_LOG
-    ASSERT(*(uint64_t *) (seg_data_dest) == SEG_MAGIC);
-    ASSERT(*(uint64_t *) (curr_src) == SEG_MAGIC);
-    curr_src += sizeof(uint64_t);
-#endif
-
-    bool        copy_all_items = false;
-    if (*cutoff_freq < 0.0001) {
-        copy_all_items = true;
-    }
-
-    int    n_scanned    = 0, n_copied = 0;
-    double mean_size    = (double) seg_src->occupied_size / seg_src->n_item;
-    double cutoff       = (1 + *cutoff_freq) / 2;
-    int    update_intvl = (int) heap.seg_size / 10;
-    int    n_th_update  = 1;
-
-    double hit;
-
-    while (curr_src - seg_data_src < offset) {
-        last_it = it;
-        it      = (struct item *) curr_src;
-
-        if (it->klen == 0 && it->vlen == 0) {
-            break;
-        }
-
-        ASSERT(seg_src->n_item >= 0);
-
-#if defined CC_ASSERT_PANIC || defined CC_ASSERT_LOG
-        ASSERT(it->magic == ITEM_MAGIC);
-#endif
-
-        it_sz = item_ntotal(it);
-        n_scanned += it_sz;
-        if (n_scanned >= n_th_update * update_intvl) {
-            n_th_update += 1;
-//            double t = (double) n_copied/n_scanned - target_ratio;
-//            if (t > 0.1 || t < -0.1) {
-            /* new change */
-            double t = (((double) n_copied) / n_scanned - target_ratio)
-                / target_ratio;
-            if (t > 0.5 || t < -0.5) {
-                cutoff = cutoff * (1 + t);
-            }
-        }
-
-        /* we will not merge a new segment, so let's copy all items left,
-         * most of the time, the impact of this is small */
-        if (!copy_all_items && (seg_dest->write_offset >= SEG_MERGE_MARGIN)
-            && curr_src - seg_data_src > SEG_MERGE_MARGIN) {
-            copy_all_items = true;
-            log_verb("seg copy %d %d/%d, last item sz %d", seg_id_src,
-                curr_src - seg_data_src,
-                seg_dest->write_offset, item_ntotal(last_it));
-        }
-
-        if (it->deleted) {
-//            hashtable_delete_it(item_key(it), it->klen, seg_id_src, curr_src - seg_data_src);
-            curr_src += it_sz;
-            continue;
-        }
-
-#ifdef USE_PRECISE_FREQ
-        hit = it->n_hit;
-#elif defined(USE_ASFC)
-        hit = hashtable_get_it_freq(item_key(it), it->klen, seg_id_src,
-            curr_src - seg_data_src);
-#elif defined(APFC_IN_OBJ)
-        hit = it->freq;
-#else
-#error need FREQ COUNTER ENABLED
-#endif
-
-#ifdef DUMP_FOR_ANALYSIS
-        int32_t t;
-        int sfc = hashtable_get_it_freq(item_key(it), it->klen, seg_id_src, curr_src - seg_data_src);
-        struct item *cit = hashtable_get_no_incr(item_key(it), it->klen, &t, NULL);
-        if (cit == NULL) {
-            fprintf(f, "%d, ", 0);
-        } else {
-            fprintf(f, "%d, ", it->n_hit);
-            if (it == cit && it->n_hit != 0)
-                ASSERT(sfc != 0);
-        }
-        fprintf(f2, "%d, ", sfc);
-#endif
-        hit = (double) hit / ((double) it_sz / mean_size);
-
-        if (hit <= cutoff && (!copy_all_items)) {
-            hashtable_evict(item_key(it), it->klen, seg_id_src,
-                curr_src - seg_data_src);
-            curr_src += it_sz;
-            continue;
-        }
-
-        if (seg_dest->write_offset + it_sz > heap.seg_size) {
-            /* TODO(jason): add a new metric */
-            if (!seg_in_full) {
-                seg_in_full = true;
-                log_debug("copy from seg %" PRId32 " to seg %" PRId32
-                    ", destination seg full %d + %d src offset %d",
-                    seg_id_src, seg_id_dest, seg_dest->write_offset, it_sz,
-                    curr_src - seg_data_src);
-            }
-
-            hashtable_evict(item_key(it), it->klen, seg_id_src,
-                curr_src - seg_data_src);
-            curr_src += it_sz;
-            continue;
-        }
-
-#ifdef REAL_COPY
-            /* first copy data */
-#ifdef USE_PMEM
-            pmem_memcpy_nodrain(seg_data_dest + seg_dest->write_offset, curr_src, it_sz);
-#else
-        memcpy(seg_data_dest + seg_dest->write_offset, curr_src, it_sz);
-#endif
-#else
-        memcpy(seg_data_dest + seg_dest->write_offset, curr_src, ITEM_HDR_SIZE + it->klen);
-#endif
-
-        /* try to relink */
-        item_up_to_date = hashtable_relink_it(item_key(it), it->klen,
-            seg_id_src, curr_src - seg_data_src, seg_id_dest,
-            seg_dest->write_offset);
-
-        if (item_up_to_date) {
-#if defined(USE_PRECISE_FREQ) || defined(DUMP_FOR_ANALYSIS)
-            ((struct item *) (seg_data_dest + seg_dest->write_offset))->n_hit = 0;
-#endif
-#ifdef APFC_IN_OBJ
-            ((struct item*)(seg_data_dest + seg_dest->write_offset))->freq = 0;
-            ((struct item*)(seg_data_dest + seg_dest->write_offset))->last_update_ts = 0;
-#endif
-            seg_dest->write_offset += it_sz;
-            seg_dest->occupied_size += it_sz;
-            seg_dest->n_item += 1;
-            seg_src->n_item -= 1;
-            n_copied += it_sz;
-        }
-
-        curr_src += it_sz;
-    }
-
-#if defined CC_ASSERT_PANIC || defined CC_ASSERT_LOG
-    if (seg_src->n_item != 0) {
-        log_warn("after copy %d items left", seg_src->n_item);
-        scan_hashtable_find_seg(seg_id_src);
-    }
-#endif
-
-#ifdef DUMP_FOR_ANALYSIS
-    fprintf(f, "\n");
-    fprintf(f2, "\n");
-    fclose(f);
-    fclose(f2);
-#endif
-    *cutoff_freq = cutoff;
-    log_debug(
-        "move items from seg %d to seg %d, new seg %d items, offset %d, cutoff %.2lf, target ratio %.2lf",
-        seg_id_src, seg_id_dest, seg_dest->n_item, seg_dest->write_offset,
-        *cutoff_freq, target_ratio);
-}
-
-bool
-seg_mergeable(struct seg *seg)
-{
-    if (seg == NULL) {
-        return false;
-    }
-    bool is_mergeable;
-    is_mergeable = seg->evictable == 1;
-    is_mergeable = is_mergeable && (seg->next_seg_id != -1);
-    /* a magic number - we don't want to merge just created seg */
-    /* TODO(jason): 600 needs to be adaptive */
-    is_mergeable =
-        is_mergeable && time_proc_sec() - seg->create_at >= min_mature_time;
-    /* don't merge segments that will expire soon */
-    is_mergeable = is_mergeable &&
-        seg->create_at + seg->ttl - time_proc_sec() > 20;
-    return is_mergeable;
-}
-
-/**
- * lock at most N_MAX_SEG_MERGE segments to prevent other threads evicting
- */
-static inline void
-prep_seg_to_merge(int32_t start_seg_id,
-                  struct seg *segs_to_merge[], int *n_seg_to_merge,
-                  double *merge_keep_ratio)
-{
-
-    *n_seg_to_merge = 0;
-    int32_t    curr_seg_id = start_seg_id;
-    struct seg *curr_seg;
-
-    uint8_t evictable;
-
-#ifdef DUMP_FOR_ANALYSIS
-    FILE *f = fopen("segInfo", "a");
-    fprintf(f, "###################################\n");
-#endif
-
-    pthread_mutex_lock(&heap.mtx);
-    for (int i = 0; i < N_MAX_SEG_MERGE; i++) {
-        if (curr_seg_id == -1) {
-            /* this could happen when prev seg is evicted */
-            break;
-        }
-        curr_seg  = &heap.segs[curr_seg_id];
-        if (!seg_mergeable(curr_seg)) {
-            curr_seg_id = curr_seg->next_seg_id;
-            continue;
-        }
-        evictable = __atomic_exchange_n(&curr_seg->evictable, 0,
-            __ATOMIC_RELAXED);
-        if (evictable == 0) {
-            /* concurrent merge and evict */
-            curr_seg_id = curr_seg->next_seg_id;
-            continue;
-        }
-        segs_to_merge[(*n_seg_to_merge)++] = curr_seg;
-        __atomic_fetch_add(&n_merge_seg, 1, __ATOMIC_RELAXED);
-        __atomic_fetch_add(&merge_seg_age_sum, proc_sec - curr_seg->create_at,
-            __ATOMIC_RELAXED);
-#ifdef DUMP_FOR_ANALYSIS
-        fprintf(f, "%d %d %d %.0lf\n", curr_seg->ttl,
-                curr_seg->merge_at == 0 ? proc_sec - curr_seg->create_at : proc_sec - curr_seg->merge_at,
-                curr_seg->n_active_byte, (double) curr_seg->occupied_size/curr_seg->n_item);
-#endif
-        curr_seg_id = curr_seg->next_seg_id;
-    }
-    pthread_mutex_unlock(&heap.mtx);
-#ifdef DUMP_FOR_ANALYSIS
-    fclose(f);
-#endif
-
-    ASSERT(*n_seg_to_merge > 1);
-}
-
-static inline void
-replace_seg_in_chain(int32_t new_seg_id, int32_t old_seg_id)
-{
-    struct seg        *new_seg = &heap.segs[new_seg_id];
-    struct seg        *old_seg = &heap.segs[old_seg_id];
-    struct ttl_bucket *tb      =
-                          &ttl_buckets[find_ttl_bucket_idx(old_seg->ttl)];
-
-    /* all modification to seg list needs to be protected by lock */
-    ASSERT(pthread_mutex_trylock(&heap.mtx) != 0);
-
-    int32_t prev_seg_id = old_seg->prev_seg_id;
-    int32_t next_seg_id = old_seg->next_seg_id;
-
-    if (prev_seg_id == -1) {
-        ASSERT(tb->first_seg_id == old_seg_id);
-
-        tb->first_seg_id = new_seg_id;
-    }
-    else {
-        heap.segs[prev_seg_id].next_seg_id = new_seg_id;
-    }
-
-    ASSERT(next_seg_id != -1);
-    heap.segs[next_seg_id].prev_seg_id = new_seg_id;
-
-    new_seg->prev_seg_id = prev_seg_id;
-    new_seg->next_seg_id = next_seg_id;
-}
-
-/* merge at most n_seg consecutive segs into one seg,
- * if the merged seg is full return earlier
- *
- * the return value indicates how many segs are merged
- *
- **/
-int32_t
-merge_segs(struct seg *segs_to_merge[], int at_most_n_seg)
-{
-    static int empty_merge      = 0;
-    static int successful_merge = 0;
-
-    int32_t    curr_seg_id;
-    struct seg *curr_seg;
-    uint8_t    accessible;
-    int        n_merged         = 0;
-
-    /* this is the next seg_id of the last segment, we keep copy of it
-     * in case there are no active objects in all these segments,
-     * this is the return value */
-    int32_t
-        last_seg_next_seg_id = segs_to_merge[at_most_n_seg - 1]->next_seg_id;
-
-    /* prepare new seg */
-    int32_t new_seg_id = seg_get_from_freepool(true);
-    seg_init(new_seg_id);
-
-    struct seg *new_seg = &heap.segs[new_seg_id];
-    ASSERT(new_seg->evictable == 0);
-
-    new_seg->create_at   = segs_to_merge[0]->create_at;
-    new_seg->merge_at    = time_proc_sec();
-    new_seg->ttl         = segs_to_merge[0]->ttl;
-    new_seg->accessible  = 1;
-    new_seg->prev_seg_id = segs_to_merge[0]->prev_seg_id;
-    double cutoff_freq = 1;
-    if (empty_merge > successful_merge && empty_merge > 2) {
-        cutoff_freq = 0;
-    }
-
-    /* start from start_seg until new_seg is full or no seg can be merged */
-    while (new_seg->write_offset < heap.seg_size * SEG_MERGE_STOP_RATIO
-        && n_merged < at_most_n_seg) {
-
-        curr_seg    = segs_to_merge[n_merged++];
-        curr_seg_id = curr_seg->seg_id;
-
-        seg_copy(new_seg_id, curr_seg_id, &cutoff_freq,
-            SEG_MERGE_TARGET_RATIO);
-        accessible = __atomic_exchange_n(&(curr_seg->accessible), 0,
-            __ATOMIC_RELAXED);
-        ASSERT(accessible == 1);
-
-        seg_wait_refcnt(curr_seg_id);
-        pthread_mutex_lock(&heap.mtx);
-        if (n_merged - 1 == 0) {
-            replace_seg_in_chain(new_seg_id, curr_seg_id);
-        }
-        else {
-            rm_seg_from_ttl_bucket(curr_seg_id);
-        }
-
-        add_seg_to_freepool(curr_seg_id, SEG_EVICTION);
-        pthread_mutex_unlock(&heap.mtx);
-    }
-
-    ASSERT(n_merged > 0);
-
-    /* if no seg has active object */
-    if (new_seg->occupied_size <= 8) {
-        new_seg->accessible = 0;
-
-        pthread_mutex_lock(&heap.mtx);
-        rm_seg_from_ttl_bucket(new_seg_id);
-        add_seg_to_freepool(new_seg_id, SEG_EVICTION);
-        pthread_mutex_unlock(&heap.mtx);
-
-        log_warn("merged %d segments with no active objects, "
-                 "return reserved seg %d", n_merged, new_seg_id);
-        for (int i = 0; i < n_merged; i++) {
-            SEG_PRINT(segs_to_merge[i]->seg_id, "seg info", log_debug);
-        }
-
-        empty_merge += 1;
-        return last_seg_next_seg_id;
-    }
-    else {
-        /* changed the status of un-merged seg */
-        for (int i = n_merged; i < at_most_n_seg; i++) {
-            uint8_t evictable = __atomic_exchange_n(
-                &segs_to_merge[i]->evictable, 1, __ATOMIC_RELAXED);
-            ASSERT(evictable == 0);
-        }
-
-        /* in seg_copy, we could copy over unused bytes */
-        memset(get_seg_data_start(new_seg_id) + new_seg->write_offset,
-            0, heap.seg_size - new_seg->write_offset);
-        new_seg->evictable = 1;
-
-        /* print stat */
-        char     merged_segs[1024];
-        int      pos       = 0;
-        for (int i         = 0; i < n_merged; i++) {
-            pos += snprintf(merged_segs + pos, 1024 - pos, "%d, ",
-                segs_to_merge[i]->seg_id);
-        }
-        log_info("ttl %d, merged %d/%d segs (%s) to seg %d, "
-                 "curr #free segs %d, new seg offset %d, occupied size %d, "
-                 "%d items",
-            new_seg->ttl, n_merged, at_most_n_seg, merged_segs, new_seg_id,
-            heap.n_free_seg, new_seg->write_offset,
-            new_seg->occupied_size, new_seg->n_item);
-        successful_merge += 1;
-    }
-
-    log_verb("***************************************************");
-    INCR_N(seg_metrics, seg_merge, n_merged);
-
-    return heap.segs[new_seg_id].next_seg_id;
-}
-//#endif
 
 static void
 heap_init(void)
@@ -1278,7 +606,7 @@ seg_heap_setup(void)
             heap.segs[i].evictable  = 0;
             heap.segs[i].accessible = 0;
 
-            add_seg_to_freepool(i, SEG_ALLOCATION);
+            seg_add_to_freepool(i, SEG_ALLOCATION);
         }
         pthread_mutex_unlock(&heap.mtx);
     }
@@ -1293,9 +621,7 @@ seg_teardown(void)
 
     stop = true;
 
-#ifdef NO_BACKGROUND_EXPIRATION
     pthread_join(bg_tid, NULL);
-#endif
 
     if (!seg_initialized) {
         log_warn("%s has never been set up", SEG_MODULE_NAME);
@@ -1349,8 +675,7 @@ seg_setup(seg_options_st *options, seg_metrics_st *metrics)
     heap.poolpath = option_str(&seg_options->datapool_path);
     heap.poolname = option_str(&seg_options->datapool_name);
 
-    heap.n_reserved_seg = 1;
-    heap.n_reserved_seg = option_uint(&seg_options->seg_n_thread);
+    heap.n_reserved_seg = 0;
 
     use_cas = option_bool(&seg_options->seg_use_cas);
 
@@ -1363,11 +688,17 @@ seg_setup(seg_options_st *options, seg_metrics_st *metrics)
 
     ttl_bucket_setup();
 
-    segevict_setup(option_uint(&options->seg_evict_opt), heap.max_nseg);
+    evict_info.merge_opt.seg_n_merge     =
+        option_uint(&seg_options->seg_n_merge);
+    evict_info.merge_opt.seg_n_max_merge =
+        option_uint(&seg_options->seg_n_max_merge);
+    segevict_setup(option_uint(&options->seg_evict_opt),
+        option_uint(&seg_options->seg_mature_time));
+    if (evict_info.policy == EVICT_MERGE_FIFO) {
+        heap.n_reserved_seg = option_uint(&seg_options->seg_n_thread);
+    }
 
-#ifndef NO_BACKGROUND_EXPIRATION
     start_background_thread(NULL);
-#endif
 
     seg_initialized = true;
