+++ src/storage/seg/segevict.c
@@ -9,8 +9,71 @@ static bool segevict_initialized;
 
 struct seg_evict_info evict_info;
 
+extern seg_metrics_st        *seg_metrics;
+extern seg_perttl_metrics_st perttl[MAX_N_TTL_BUCKET];
+
+
 #define IS_BAD_EVICT_CAN(seg)                                     \
     (seg->w_refcount > 0 || seg->evictable == 0 || (seg->create_at + seg->ttl) - time_proc_sec() < 5 || seg->next_seg_id == -1)
+
+
+/**
+ * find the least valuable segment in DRAM
+ * return seg_id
+ */
+evict_rstatus_e
+least_valuable_seg(int32_t *seg_id);
+
+
+bool
+seg_evictable(struct seg *seg)
+{
+    if (seg == NULL) {
+        return false;
+    }
+
+    bool is_evictable;
+    /* although we check evictable here, we will check again after
+     * we grab the lock, so this is part of the opportunistic concurrency control */
+    is_evictable = (seg->evictable == 1) && (seg->next_seg_id != -1);
+
+    /* a magic number - we don't want to merge just created seg */
+    /* TODO(jason): the time needs to be adaptive */
+    is_evictable =
+        is_evictable
+            && (time_proc_sec() - seg->create_at >= evict_info.seg_mature_time);
+
+    /* don't merge segments that will expire soon */
+    is_evictable = is_evictable &&
+        seg->create_at + seg->ttl - time_proc_sec() > 20;
+
+    return is_evictable;
+}
+
+evict_rstatus_e
+seg_evict(int32_t *evicted_seg_id) {
+    evict_rstatus_e status;
+
+    status = least_valuable_seg(evicted_seg_id);
+    if (status == EVICT_NO_AVAILABLE_SEG) {
+        log_warn("unable to find seg to evict");
+        INCR(seg_metrics, seg_evict_ex);
+
+        return EVICT_NO_AVAILABLE_SEG;
+    }
+
+    log_verb("evict segment %"PRId32, *evicted_seg_id);
+
+    if (rm_all_item_on_seg(*evicted_seg_id, SEG_EVICTION)) {
+        INCR(seg_metrics, seg_evict);
+
+        return EVICT_OK;
+    }
+
+    *evicted_seg_id = -1;
+    return EVICT_CANNOT_LOCK_SEG;
+}
+
 
 /* maybe we should use # of req instead of real time to make decision */
 static inline bool
@@ -22,10 +85,11 @@ should_rerank(void)
     prev_sec = __atomic_load_n(&evict_info.last_update_time, __ATOMIC_SEQ_CST);
     rerank   = prev_sec == -1 || curr_sec - prev_sec > UPDATE_INTERVAL;
 
-    rerank = rerank || evict_info.nseg - evict_info.idx_rseg < 8;
+    rerank = rerank || heap.max_nseg - evict_info.idx_rseg < 8;
 
     if (rerank) {
-        __atomic_store_n(&evict_info.last_update_time, curr_sec, __ATOMIC_SEQ_CST);
+        __atomic_store_n(&evict_info.last_update_time, curr_sec,
+            __ATOMIC_SEQ_CST);
     }
 
     return rerank;
@@ -105,13 +169,6 @@ cmp_seg_smart(const void *d1, const void *d2)
 #endif
 }
 
-// static inline void
-//_predict_n_hit(void)
-//{
-//    ;
-//}
-
-
 static inline void
 rank_seg(void)
 {
@@ -137,7 +194,7 @@ rank_seg(void)
     }
 
     ASSERT(cmp != NULL);
-    qsort(evict_info.ranked_seg_id, evict_info.nseg, sizeof(uint32_t), cmp);
+    qsort(evict_info.ranked_seg_id, heap.max_nseg, sizeof(uint32_t), cmp);
 
 
 //#ifdef TRACK_ADVANCED_STAT
@@ -182,7 +239,7 @@ least_valuable_seg(int32_t *seg_id)
         }
         if (i == heap.max_nseg) {
             log_warn("unable to find a segment to evict");
-            return EVICT_NO_SEALED_SEG;
+            return EVICT_NO_AVAILABLE_SEG;
         }
         else {
 //            log_warn("%d %d %d %d", *seg_id, seg->seg_id, seg->next_seg_id, i);
@@ -200,11 +257,11 @@ least_valuable_seg(int32_t *seg_id)
 
         /* it is OK if we read a staled seg.sealed because we will double check
          * when we perform real eviction */
-        while (IS_BAD_EVICT_CAN(seg) && evict_info.idx_rseg < evict_info.nseg) {
+        while (IS_BAD_EVICT_CAN(seg) && evict_info.idx_rseg < heap.max_nseg) {
             *seg_id = evict_info.ranked_seg_id[evict_info.idx_rseg++];
             seg = &heap.segs[*seg_id];
         }
-        if (evict_info.idx_rseg >= evict_info.nseg) {
+        if (evict_info.idx_rseg >= heap.max_nseg) {
             seg = &heap.segs[evict_info.ranked_seg_id[0]];
             log_warn("unable to find a segment to evict, top seg %d, ttl %d, "
                      "accessible %d evictable %d, age %d, w_ref %d, next_seg %d",
@@ -214,7 +271,7 @@ least_valuable_seg(int32_t *seg_id)
             evict_info.idx_rseg = 0;
             /* better return a less utilized one */
             pthread_mutex_unlock(&evict_info.mtx);
-            return EVICT_NO_SEALED_SEG;
+            return EVICT_NO_AVAILABLE_SEG;
         }
 
         pthread_mutex_unlock(&evict_info.mtx);
@@ -240,25 +297,33 @@ segevict_teardown(void)
 }
 
 void
-segevict_setup(evict_policy_e ev_policy, int32_t nseg)
+segevict_setup(evict_policy_e ev_policy, uintmax_t seg_mature_time)
 {
     uint32_t i = 0;
 
     if (segevict_initialized) {
         log_warn("segevict has already initialized");
+
         segevict_teardown();
     }
 
     evict_info.last_update_time = -1;
     evict_info.policy           = ev_policy;
-    evict_info.nseg             = nseg;
-    evict_info.ranked_seg_id    = cc_zalloc(sizeof(int32_t) * nseg);
+    evict_info.ranked_seg_id    = cc_zalloc(sizeof(int32_t) * heap.max_nseg);
     evict_info.idx_rseg         = 0;
     pthread_mutex_init(&evict_info.mtx, NULL);
 
-    for (i = 0; i < nseg; i++) {
+    for (i = 0; i < heap.max_nseg; i++) {
         evict_info.ranked_seg_id[i] = i;
     }
+
+    /* initialize merged-based eviction policy */
+    struct merge_opts *mopt = &evict_info.merge_opt;
+    mopt->target_ratio = 1.0 / mopt->seg_n_merge;
+    /* stop if the bytes on the merged seg is more than the threshold */
+    mopt->stop_ratio   = mopt->target_ratio * (mopt->seg_n_merge - 1) + 0.05;
+    mopt->stop_bytes   = (int32_t) (heap.seg_size * mopt->stop_ratio);
+
 
     srand(time(NULL));
     segevict_initialized = true;
