+++ src/storage/seg/segevict.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "constant.h"
+#include "segmerge.h"
 
 #include <time/time.h>
 #include <cc_mm.h>
@@ -20,7 +21,8 @@ typedef enum {
 
 typedef enum evict_rstatus {
     EVICT_OK,
-    EVICT_NO_SEALED_SEG,
+    EVICT_CANNOT_LOCK_SEG,
+    EVICT_NO_AVAILABLE_SEG,
 
     EVICT_OTHER,
 } evict_rstatus_e;
@@ -28,25 +30,41 @@ typedef enum evict_rstatus {
 
 struct seg_evict_info {
     evict_policy_e      policy;
+
+    struct merge_opts    merge_opt;
+
+    /* segment younger than seg_mature_time should not be selected */
+    int32_t             seg_mature_time;
+
     proc_time_i         last_update_time;
-    int32_t             nseg;
+
     int32_t             *ranked_seg_id;  /* ranked seg ids from the least
                                           * valuable to the most valuable */
     int32_t             idx_rseg;        /* curr index in ranked seg id array */
+
     pthread_mutex_t     mtx;
 };
 
 
+struct seg;
 /**
- * find the least valuable segment in DRAM
- * return seg_id
+ * check whether a segment can be evicted,
+ * a segment cannot be evicted if
+ * 1. it is expired or expiring soon
+ * 2. is is being evicted by another thread
+ * 3. it is the last segment of the chain (active be written to)
+ * 4. it is too young (age smaller than seg_mature_time)
  */
+bool
+seg_evictable(struct seg *seg);
+
 evict_rstatus_e
-least_valuable_seg(int32_t *seg_id);
+seg_evict(int32_t *evicted_seg_id);
 
 
-/* this must be called after seg_setup has finished */
-void segevict_setup(evict_policy_e ev_policy, int32_t nseg);
+void
+segevict_setup(evict_policy_e ev_policy, uintmax_t seg_mature_time);
 
-void segevict_teardown(void);
+void
+segevict_teardown(void);
